fix: str→int cast in achievement checks — fixes '360' str error

+++ backend/app/services/token_billing.py
@@ -115,8 +115,8 @@ async def deduct_balance(db: AsyncSession, tg_id: int, amount: float) -> dict:
     if actual_deduct > 0:
         import asyncio
         from app.routers.achievements import check_and_update
-        total_spent_rub = round((float(user.total_deposited_usd or 0) - float(user.balance_usd or 0)) * 95)
-        asyncio.create_task(check_and_update(tg_id, "spent_rub", max(0, total_spent_rub)))
+        total_spent_rub = int(round((float(user.total_deposited_usd or 0) - float(user.balance_usd or 0)) * 95))
+        asyncio.create_task(check_and_update(int(tg_id), "spent_rub", max(0, total_spent_rub)))
 
     return {
         "success": True,
